Cleanup consensus to be grouped correctly
