+++ technical/consensus/__init__.py
@@ -1,0 +1,4 @@
+# flake8: noqa: F401
+from technical.consensus.consensus import Consensus
+from technical.consensus.movingaverage import MovingAverageConsensus
+from technical.consensus.oscillator import OscillatorConsensus
